requirements.txt
langchain
pandas
flask
plotly
serpapi
python-dotenv
langchain_groq
gunicorn
